+++ GRDB/Core/Statement.swift
@@ -38,7 +38,7 @@ public class Statement {
     }
     
     private init(database: Database, sql: String, observer: StatementCompilationObserver) throws {
-        database.preconditionValidQueue()
+        DatabaseScheduler.preconditionValidQueue(database)
         
         observer.start()
         defer { observer.stop() }
@@ -326,7 +326,7 @@ public struct DatabaseSequence<Element>: SequenceType {
     private init(statement: SelectStatement, element: () -> Element) {
         self.generateImpl = {
             // Check that generator is built on a valid queue.
-            statement.database.preconditionValidQueue("Database was not used on the correct thread. Iterate sequences in a protected dispatch queue, or consider using an array returned by fetchAll() instead.")
+            DatabaseScheduler.preconditionValidQueue(statement.database, "Database was not used on the correct thread. Iterate sequences in a protected dispatch queue, or consider using an array returned by fetchAll() instead.")
             
             // Support multiple sequence iterations
             try statement.reset()
@@ -353,7 +353,7 @@ public struct DatabaseSequence<Element>: SequenceType {
         // to be used on the database queue.
         return DatabaseSequence() {
             // Check that generator is built on a valid queue.
-            database.preconditionValidQueue("Database was not used on the correct thread. Iterate sequences in a protected dispatch queue, or consider using an array returned by fetchAll() instead.")
+            DatabaseScheduler.preconditionValidQueue(database, "Database was not used on the correct thread. Iterate sequences in a protected dispatch queue, or consider using an array returned by fetchAll() instead.")
             return DatabaseGenerator()
         }
     }
@@ -438,7 +438,7 @@ public final class UpdateStatement : Statement {
     /// - parameter arguments: Statement arguments.
     /// - throws: A DatabaseError whenever an SQLite error occurs.
     public func execute(arguments arguments: StatementArguments? = nil) throws {
-        database.preconditionValidQueue()
+        DatabaseScheduler.preconditionValidQueue(database)
         
         // Force arguments validity. See SelectStatement.fetchSequence(), and Database.execute()
         try! prepareWithArguments(arguments)