+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Release Notes
 
 - The Record class no longer adopts CustomStringConvertible. This frees the `description` identifier for your record properties. Fixes [#58](https://github.com/groue/GRDB.swift/issues/58).
 
+- Several database connections can now be used at the same time: you can move values from one database to another. Fixes [#55](https://github.com/groue/GRDB.swift/issues/55).
+
+
 **Breaking Changes**
 
 - The maximum number of reader connections in a database pool is now configured in a Configuration object.